+++ src/mindlessgen/prog/config.py
@@ -347,7 +347,7 @@ class GenerateConfig(BaseConfig):
                         tmp[key] = composition[key]
                     else:
                         raise KeyError(
-                            f"Element with atomic number {key+1} (provided key: {key}) not found in the periodic table."
+                            f"Element with atomic number {key + 1} (provided key: {key}) not found in the periodic table."
                         )
             self._element_composition = tmp
             return
@@ -723,7 +723,7 @@ class PostProcessConfig(BaseConfig):
         """
         if not isinstance(engine, str):
             raise TypeError("Postprocess engine should be a string.")
-        if engine not in ["xtb", "orca", "gp3", "turbomole"]:
+        if engine not in ["xtb", "orca", "gxtb", "turbomole"]:
             raise ValueError("Postprocess engine can only be xtb, orca or turbomole.")
         self._engine = engine
 
@@ -1006,6 +1006,53 @@ class TURBOMOLEConfig(BaseConfig):
         self._scf_cycles = max_scf_cycles
 
 
+class GXTBConfig(BaseConfig):
+    """
+    Configuration class for g-xTB.
+    """
+
+    def __init__(self: GXTBConfig) -> None:
+        self._gxtb_path: str | Path = "gxtb"
+        self._scf_cycles: int = 100
+
+    def get_identifier(self) -> str:
+        return "gxtb"
+
+    @property
+    def gxtb_path(self):
+        """
+        Get the g-xTB path.
+        """
+        return self._gxtb_path
+
+    @gxtb_path.setter
+    def gxtb_path(self, gxtb_path: str | Path):
+        """
+        Set the g-xTB path.
+        """
+        if not isinstance(gxtb_path, str | Path):
+            raise TypeError("gxtb_path should be a string or Path.")
+        self._gxtb_path = gxtb_path
+
+    @property
+    def scf_cycles(self):
+        """
+        Get the maximum number of SCF cycles.
+        """
+        return self._scf_cycles
+
+    @scf_cycles.setter
+    def scf_cycles(self, max_scf_cycles: int):
+        """
+        Set the maximum number of SCF cycles.
+        """
+        if not isinstance(max_scf_cycles, int):
+            raise TypeError("Max SCF cycles should be an integer.")
+        if max_scf_cycles < 1:
+            raise ValueError("Max SCF cycles should be greater than 0.")
+        self._scf_cycles = max_scf_cycles
+
+
 class ConfigManager:
     """
     Overall configuration manager for the program.
@@ -1022,6 +1069,7 @@ class ConfigManager:
         self.refine = RefineConfig()
         self.postprocess = PostProcessConfig()
         self.generate = GenerateConfig()
+        self.gxtb = GXTBConfig()
 
         if config_file:
             self.load_from_toml(config_file)