+++ src/mindlessgen/qm/__init__.py
@@ -5,8 +5,8 @@ This module contains all QM-related functions and classes.
 from .base import QMMethod
 from .xtb import XTB, get_xtb_path
 from .orca import ORCA, get_orca_path
-from .gp3 import GP3, get_gp3_path
 from .tm import Turbomole, get_turbomole_path
+from .gxtb import GXTB, get_gxtb_path
 
 __all__ = [
     "XTB",
@@ -14,8 +14,8 @@ __all__ = [
     "QMMethod",
     "ORCA",
     "get_orca_path",
-    "GP3",
-    "get_gp3_path",
     "Turbomole",
     "get_turbomole_path",
+    "GXTB",
+    "get_gxtb_path",
 ]